+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 All notable changes to the Tax Switch for WooCommerce plugin will be documented in this file.
 
 ## [1.5.9]BETA
+### Added
+- Compatibility for Kapee theme (version ≥ 1.6.20)
 ### Removed
-Load_plugin_textdomain() because it has been discouraged since WordPress version 4.6.
+- Load_plugin_textdomain() because it has been discouraged since WordPress version 4.6.
 
 ## [1.5.8] - 2025-07-04
 ### Added